refactor(#21): quadratic equations were renamed for issues
